refactor: conan lock: working on extractor
