+++ detectable/src/main/java/com/synopsys/integration/detectable/detectables/conan/lockfile/ConanLockfileDetectable.java
@@ -97,7 +97,7 @@ public class ConanLockfileDetectable extends Detectable {
     @Override
     public Extraction extract(ExtractionEnvironment extractionEnvironment) throws ExecutableFailedException {
         logger.info("*** CONAN LOCKFILE extract() called.");
-        conanLockfileExtractor.extract(lockfile);
+        Extraction extract = conanLockfileExtractor.extract(lockfile, conanLockfileExtractorOptions);
         return new Extraction.Builder().failure("Conan lockfile extractor not yet implemented").build();
     }
 }
